(윤수빈) refactor: 사용자의 gameData 데이터를 패킷에 맞게 생성하는 함수 파일 분리
